Refactor typst-svg
